Clean up.

+++ manager/src/main/scala/org/broadinstitute/transporter/web/Page.scala
@@ -3,7 +3,13 @@ package org.broadinstitute.transporter.web
 import io.circe.{Decoder, Encoder}
 import io.circe.derivation.{deriveDecoder, deriveEncoder}
 
-case class Page[I](items: List[I], total: Long)
+/**
+  * Generic API envelope for a page of query results.
+  *
+  * @param items query results to return to the client
+  * @param total the number of elements included in `items`
+  */
+case class Page[I](items: List[I], total: Int)
 
 object Page {
   implicit def decoder[I: Decoder]: Decoder[Page[I]] = deriveDecoder
